+++ ddb/feature/docker/schema.py
@@ -2,6 +2,7 @@
 
 from marshmallow import fields, Schema
 
+from ddb.config import config
 from ddb.feature.schema import FeatureSchema
 
 
@@ -47,6 +48,12 @@ class DebugSchema(Schema):
     host = fields.String(required=True, default=None)  # default is set in feature _configure_defaults
 
 
+def _default_restart_policy():
+    if config.data.get("core.env.current") == "dev":
+        return "no"
+    return "unless-stopped"
+
+
 class DockerSchema(FeatureSchema):
     """
     Docker schema.
@@ -54,7 +61,7 @@ class DockerSchema(FeatureSchema):
     user = fields.Nested(UserSchema(), required=True, default=UserSchema())
     ip = fields.String(required=True, default=None)  # default is set in feature _configure_defaults
     debug = fields.Nested(DebugSchema(), default=DebugSchema())
-    restart_policy = fields.String(required=True, default="no")
+    restart_policy = fields.String(required=True, default=_default_restart_policy)
     port_prefix = fields.Integer(required=False)  # default is set in feature _configure_defaults
     registry = fields.Nested(RegistrySchema(), required=True, default=RegistrySchema())
     interface = fields.String(required=True, default="docker0")